feat: 사용자 권한 검사 및 에러 처리 로직 추가
- 사용자 권한 검사를 위한 사용자 정의 어노테이션(@Login) 및 Argument Resolver 구현
- JWT 에 사용자 식별용 userSeq 정보 추가하여 인증 및 권한 확인 처리
- API 사용 중 에러 응답 처리를 위한 에러 코드 추가 (UNAUTHORIZED_ACTION 등)
- 사용자 자신의 정보만 수정/삭제할 수 있도록 컨트롤러에 검사 로직 적용

+++ userService/src/main/java/ecommerce/userService/user/controller/UserController.java
@@ -1,5 +1,9 @@
 package ecommerce.userService.user.controller;
 
+import static ecommerce.userService.exception.ErrorCodes.*;
+
+import ecommerce.userService.global.Login;
+import ecommerce.userService.global.LoginUser;
 import ecommerce.userService.user.dto.UserRequest;
 import ecommerce.userService.user.dto.UserResponse;
 import ecommerce.userService.user.dto.UserUpdateRequest;
@@ -37,13 +41,19 @@ public class UserController {
 	}
 
 	@PutMapping("/{id}")
-	public ApiResult<UserResponse> updateUser(@PathVariable Long id,
+	public ApiResult<UserResponse> updateUser(@Login LoginUser user, @PathVariable Long id,
 		@Valid @RequestBody UserUpdateRequest userRequest) {
+		if (!user.getUserSeq().equals(id)) {
+			return ApiResult.failure(UNAUTHORIZED_ACTION, "Not authorized");
+		}
 		return ApiResult.success(userService.updateUser(id, userRequest));
 	}
 
 	@DeleteMapping("/{id}")
-	public ApiResult<Void> deleteUser(@PathVariable Long id) {
+	public ApiResult<Void> deleteUser(@Login LoginUser user, @PathVariable Long id) {
+		if (!user.getUserSeq().equals(id)) {
+			return ApiResult.failure(UNAUTHORIZED_ACTION, "Not authorized");
+		}
 		userService.deleteUser(id);
 		return ApiResult.success(null);
 	}
